+++ documentation/changelog.md
@@ -62,6 +62,7 @@ description: List of all the changes and current versions.
 + Added Locales for Vehicle Status Server Side
 + Added Locales for Clothing Server Side
 + Added Locales for ATM Server Side
++ Added Named Interactions for Client Side
 ```
 
 ## 1.8.0 - February 28-x, 2021